Refactor(blacklist): 컴포넌트 분리
